+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.11.55](https://github.com/jensvogt/awsmock-qt-ui/compare/v1.11.54...v1.11.55) (2026-05-04)
+
+
+### Bug Fixes
+
+* resend SNS messages ([b26548d](https://github.com/jensvogt/awsmock-qt-ui/commit/b26548d0510def37f39c8512b402feaadf8438b6))
+
+## [1.11.54](https://github.com/jensvogt/awsmock-qt-ui/compare/1.11.53...v1.11.54) (2026-05-04)
+
+
+### Bug Fixes
+
+* resend SNS messages ([7cb0476](https://github.com/jensvogt/awsmock-qt-ui/commit/7cb0476d26ec24dfc6f50d44a6efc7ffe7ab199c))
+
 ## [1.11.53](https://github.com/jensvogt/awsmock-qt-ui/compare/1.11.52...v1.11.53) (2026-05-04)